Revise Platformatic DB Authorization docs
I've rewritten the 'Authorization' section of the Platformatic DB
'Configuration' docs to be in the same format as the other sections
on that page. I've moved most of the documentation that was originally
on that page into pages under the Platformatic DB Authorization category.

I've also made lots of minor edits in an attemp to clarify how
authorization works in Platformatic DB.

Signed-off-by: Simon Plenderleith <simon@simonplend.co.uk>