reverse dependency between runtime and start
Prior to this change, the runtime depended on the start API.
This lead to a few problems:

- Due to would be circular dependencies, start could not depend
  on runtime, so runtime apps could not be started via the
  universal start command.
- The previous point propagated to consumers of start who needed
  to add code to special case the runtime.

This commit moves the unified API code into the runtime and
updates it so that it works with runtime apps. The start module
is now just a wrapper around the API exported from the runtime.