runtime: add _write() to MessagePortWritable
In some situations, such as when DB migrations fail during
startup, the error logs were not being handled properly. According
to the Node.js docs, custom writables such as MessagePortWritable
need to implement one or both of _write() and _writev(). The
runtime only implemented _writev(), and this seemed to be the
cause of the missing logs. This commit adds _write(), and seems
to resolve the missing logs issue.

Fixes: https://github.com/platformatic/dashformatic/issues/648