db: fix consistency to reflect the change from migrationsTable to table (#2171)
* 📝 docs(db): update configuration.md and config.d.ts to reflect the change from migrationsTable to table
💡 refactor(schema.js): remove migrationsTable from the migrations object and add table with description and default value
The `migrationsTable` property in the `PlatformaticDB` interface and the `migrationsTable` field in the `migrations` object have been renamed to `table` to improve consistency and clarity. The `table` property now represents the table created to track the schema version. The documentation and type definitions have been updated accordingly.

* 📝 docs(db): update configuration.md and config.d.ts to reflect the change from migrationsTable to table
💡 refactor(schema.js): remove migrationsTable from the migrations object and add table with description and default value
The `migrationsTable` property in the `PlatformaticDB` interface and the `migrationsTable` field in the `migrations` object have been renamed to `table` to improve consistency and clarity. The `table` property now represents the table created to track the schema version. The documentation and type definitions have been updated accordingly.

Signed-off-by: Mirko Pira <mp@mirkodev.com>

---------

Signed-off-by: Mirko Pira <mp@mirkodev.com>
Co-authored-by: Mirko Pira <mp@mirkodev.com>