fix(sql-mapper/types): enhance null value type support and documentation (#2360)
* ♻️ (mapper.d.ts): update WhereCondition interface to allow null values for eq and neq properties
The WhereCondition interface has been updated to allow null values for the eq and neq properties. This change improves the flexibility of the interface by allowing these properties to be explicitly set to null when needed.

Signed-off-by: Mirko Pira <mp@mirkodev.com>

* 📝 (api.md): Add explanation on how `eq` and `neq` operators handle null values in SQL queries
✨ (api.md): Improve documentation on handling null values when using `eq` and `neq` operators in SQL queries
The documentation now includes a section explaining how the `eq` and `neq` operators handle null values in SQL queries. This addition enhances the clarity and understanding of how these operators behave when dealing with null values, providing users with a better reference for utilizing them in their queries.

Signed-off-by: Mirko Pira <mp@mirkodev.com>

* 📝 (api.md): Remove redundant explanation

Signed-off-by: Mirko Pira <mp@mirkodev.com>

* ✨ (mapper.test-d.ts): Add assertions for find where condition with null in mapper tests

Signed-off-by: Mirko Pira <mp@mirkodev.com>

---------

Signed-off-by: Mirko Pira <mp@mirkodev.com>
Co-authored-by: Mirko Pira <mp@mirkodev.com>