fix(sql-mapper/types): add missing UpdateMany interface (#2361)
* ✨ (mapper.d.ts): introduce a new interface 'UpdateMany' to support updating multiple entity rows in the database
The 'UpdateMany' interface is added to the mapper.d.ts file to provide a way to update multiple entity rows in the database. This new interface includes options for specifying the where condition, entity fields to update, fields to be returned, and transaction participation. This enhancement allows for more flexibility and control when updating multiple entity rows in the database.

Signed-off-by: Mirko Pira <mp@mirkodev.com>

* ✨ (mapper.test-d.ts): Add assertions for updateMany method in mapper tests

Signed-off-by: Mirko Pira <mp@mirkodev.com>

---------

Signed-off-by: Mirko Pira <mp@mirkodev.com>
Co-authored-by: Mirko Pira <mp@mirkodev.com>
Co-authored-by: Matteo Collina <hello@matteocollina.com>