🐛 fix: exclude platformatic-admin role when other roles are present in findRule function
Add constant PLT_ADMIN_ROLE for 'platformatic-admin' role and update findRule to
filter out platformatic-admin rules if other roles exist. This change enables user
impersonation by allowing admins to test with the exact permissions of other users
instead of always applying the admin's full permissions.

Add tests to verify that platformatic-admin is excluded when other roles are present
and included when it is the only role.

Signed-off-by: Mirko Pira <mp@mirkodev.com>